requirements.txt
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
cleanco==2.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
#cupy==12.2.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.105.0 ; python_version >= "3.11" and python_version < "4.0"
fastrlock==0.8.2 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.0.1 ; python_version >= "3.11" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
joblib==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
libretranslatepy==2.1.1 ; python_version >= "3.11" and python_version < "4.0"
llvmlite==0.41.1 ; python_version >= "3.11" and python_version < "4.0"
lxml==4.9.3 ; python_version >= "3.11" and python_version < "4.0"
name-matching==0.8.9 ; python_version >= "3.11" and python_version < "4.0"
#numba==0.58.1 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.2 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.14.5 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.5.2 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
scikit-learn==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
scipy==1.11.4 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
sparse-dot-topn==0.3.6 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.23 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
threadpoolctl==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.11" and python_version < "4.0"
translate==3.6.1 ; python_version >= "3.11" and python_version < "4.0"
transliterate==1.10.2 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
wget==3.2 ; python_version >= "3.11" and python_version < "4.0"

uvicorn[standard]
Jinja2
python-multipart